llvmPackages.stdenv: treat Darwin as a “system” libc++ platform
Applying `overrideLibcxx` by default allows various LLVM versions to use
the same underlying libc++ on Darwin. This works because libc++ is built
by default with the stable ABI. Using the same libc++ regardless of LLVM
version avoids potential ODR violations from linking multiple libc++
versions in the same process.

This is similar to the default on Linux, which uses libstdc++ instead of
libc++. On both platforms, to force a stdenv to link against that LLVM
version’s libc++, use `libcxxStdenv`.